requirements.txt
pandas
numpy
ccxt
python-telegram-bot
joblib
scikit-learn
matplotlib
sqlalchemy
python-dotenv
requests
tensorflow
